Updated with support for requestAll style APIs to receive an initial page or page number to request.
This requires observer use, and the observer must have a variable added to its object called "initPage". This can be set to the page number only, or it can be set to the full URL you wish to start your API requests from.

Helpful when collecting large amounts of data and you may need to resume your collection by picking up the last page number.